backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
google-generativeai
python-multipart
pydantic-settings
python-dotenv
